+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - `string.encodeBase64`
     - `string.decodeBase64`
     - `string.isdigit`
+    - `string.isLowerCase`
 ## 1.0.0
 - Add:
     - `string.toTitleCase`